removed unnecessary Cast and longValue()

+++ nanopay/src/net/nanopay/account/AuthenticatedBalanceDAO.java
@@ -48,7 +48,7 @@ public class AuthenticatedBalanceDAO
     }
 
     Account account = balance.findAccount(x);
-    if ( user.getId() != ((Long)account.getOwner()).longValue()) {
+    if ( user.getId() != account.getOwner()) {
       // TODO/REVIEW: Reassign user's accounts?
       // We'll end up with multiple accounts for same currency owned by admin
     }
@@ -69,7 +69,7 @@ public class AuthenticatedBalanceDAO
     // fetch balance from delegate and verify user either owns the balance or has global read access
     if ( balance != null ) {
       Account account = balance.findAccount(x);
-      if ( user.getId() != ((Long)account.getOwner()).longValue() && ! auth.check(x, GLOBAL_BALANCE_READ) ) {
+      if ( user.getId() != account.getOwner() && ! auth.check(x, GLOBAL_BALANCE_READ) ) {
         return null;
       }
     }
@@ -109,7 +109,7 @@ public class AuthenticatedBalanceDAO
 
     if ( balance != null ) {
       Account account = balance.findAccount(x);
-      if ( user.getId() != ((Long)account.getOwner()).longValue() && ! auth.check(x, GLOBAL_BALANCE_DELETE) ) {
+      if ( user.getId() != account.getOwner() && ! auth.check(x, GLOBAL_BALANCE_DELETE) ) {
         throw new RuntimeException("Unable to delete balance.");
       }
     }
